+++ data/core.lua
@@ -1,176 +1,97 @@
 --=====================================================================================
---=====================================================================================
-RU = LibStub("AceAddon-3.0"):NewAddon("RU", "AceEvent-3.0", "AceConsole-3.0")
---=====================================================================================
---=====================================================================================
-local AC = LibStub("AceConfig-3.0")
-local ACD = LibStub("AceConfigDialog-3.0")
---=====================================================================================
---=====================================================================================
-local functionsHalted = false
---=====================================================================================
---=====================================================================================
-function RU:OnInitialize()
-    self.db = LibStub("AceDB-3.0"):New("RUDB", self.defaults, true)
-    AC:RegisterOptionsTable("RU_Options", self.options)
-    self.optionsFrame = ACD:AddToBlizOptions("RU_Options", "Rep Up!")
-    profiles = LibStub("AceDBOptions-3.0"):GetOptionsTable(self.db)
-    AC:RegisterOptionsTable("RU_Profiles", profiles)
-    ACD:AddToBlizOptions("RU_Profiles", "Profiles", "Rep Up!")
-    self:RegisterChatCommand("repup", "SlashCommand")
-    self:RegisterChatCommand("rep", "SlashCommand")
-    self:RegisterChatCommand("ru", "SlashCommand")
-end
---=====================================================================================
+-- Event Registration
 --=====================================================================================
 function RU:OnEnable()
     self:RegisterEvent("MAJOR_FACTION_RENOWN_LEVEL_CHANGED")
     self:RegisterEvent("PLAYER_ENTERING_WORLD")
-    self:RegisterEvent("UPDATE_FACTION")
+    self:ReputationChatFrameHook() -- Hook the chat frame here instead of using UPDATE_FACTION
 end
+
 --=====================================================================================
+-- Utility Functions
+--=====================================================================================
+local muteSoundIDs = {12345, 67890} -- Example sound IDs to mute
+
+-- Function to mute specific sounds by their sound IDs
+function RU:MuteSounds()
+    for _, soundID in ipairs(muteSoundIDs) do
+        MuteSoundFile(soundID) 
+        if debugMode then
+            self:DebugMessage("[|cff3bbc00RU|r] [|cff808080DEBUG|r] Muting sound with ID: |cff8080ff" .. soundID .. "|r")
+        end
+    end
+end
+
+--=====================================================================================
+-- Event Handlers
 --=====================================================================================
 function RU:PLAYER_ENTERING_WORLD(...)
-    C_Timer.After(1, function()
+    if debugMode then
+        self:DebugMessage("[|cff3bbc00RU|r] [|cff808080DEBUG|r] |cffc586c0PLAYER_ENTERING_WORLD|r event triggered.")
+    end
+    C_Timer.After(15, function()
         functionsHalted = false
     end)
     functionsHalted = true
     self:MuteSounds()
 end
---=====================================================================================
---=====================================================================================
-local muteSoundIDs = {
-    568016,
-    2066672
-}
---=====================================================================================
---=====================================================================================
-function RU:MuteSounds()
-    for _, soundID in ipairs(muteSoundIDs) do
-        MuteSoundFile(soundID)
-    end
-end
---=====================================================================================
---=====================================================================================
-local function RandomSoundID()
-    local validSoundIDs = {}
-    for soundID, soundList in pairs(sounds) do
-        for _, _ in pairs(soundList) do
-            table.insert(validSoundIDs, {table = sounds, id = soundID})
-        end
-    end
-    for soundID, soundList in pairs(defaultSounds) do
-        for _, _ in pairs(soundList) do
-            table.insert(validSoundIDs, {table = defaultSounds, id = soundID})
-        end
-    end
-    if #validSoundIDs == 0 then
-        return nil
-    end
-    local randomIndex = math.random(1, #validSoundIDs)
-    return validSoundIDs[randomIndex]
-end
---=====================================================================================
---=====================================================================================
-local function SelectSound(soundID)
-    if not soundID or soundID == 2 then
-        return RandomSoundID()
-    end
-    return {table = sounds, id = soundID}
-end
---=====================================================================================
---=====================================================================================
-function RU:MAJOR_FACTION_RENOWN_LEVEL_CHANGED(event, ...)
+
+function RU:MAJOR_FACTION_RENOWN_LEVEL_CHANGED()
     if functionsHalted then return end
+    if debugMode then
+        self:DebugMessage("[|cff3bbc00RU|r] [|cff808080DEBUG|r] |cffc586c0MAJOR_FACTION_RENOWN_LEVEL_CHANGED|r event triggered.")
+    end
     local sound = SelectSound(RU.db.profile.RenownSoundSelect)
     local volumeLevel = RU.db.profile.RenownVolume
-    local soundFile
-    if volumeLevel == 0 then
-        return
-    end
-    if sound.id == 1 then
-        soundFile = defaultSounds[5][volumeLevel]
-    else
-        soundFile = sound.table[sound.id][volumeLevel]
-    end
-    if soundFile then
-        PlaySoundFile(soundFile, "MASTER")
-    end
+    PlaySelectedSound(sound, volumeLevel, defaultSounds[5])
 end
---===============================================
-function TestRenownSound()
-    local sound = SelectSound(RU.db.profile.RenownSoundSelect)
-    local volumeLevel = RU.db.profile.RenownVolume
-    local soundFile
-    if volumeLevel == 0 then
-        return
-    end
-    if sound.id == 1 then
-        soundFile = defaultSounds[5][volumeLevel]
-    else
-        soundFile = sound.table[sound.id][volumeLevel]
-    end
-    if soundFile then
-        PlaySoundFile(soundFile, "MASTER")
-    end
-end
---=====================================================================================
---=====================================================================================
-local TrackedFactions = {}
---===============================================
-function RU:UPDATE_FACTION(event, ...)
+
+function RU:ReputationRankIncrease(factionName, newRank)
     if functionsHalted then return end
-    for i = 1, GetNumFactions() do
-        local _, _, newstanding, _, _, _, _, _, isheader, _, hasrep, _, _, faction = GetFactionInfo(i)
-        if not faction then return nil end
-        if (not isheader or hasrep) and (newstanding or 0) > 0 then
-            local oldstanding = TrackedFactions[faction]
-            if oldstanding and oldstanding < newstanding and RU.db.profile.RepSoundSelect then
-                local sound = SelectSound(RU.db.profile.RepSoundSelect)
-                local volumeLevel = RU.db.profile.RepVolume
-                local soundFile
-                if volumeLevel == 0 then
-                    return
-                end
-                if sound.id == 1 then
-                    soundFile = defaultSounds[6][volumeLevel]
-                else
-                    soundFile = sound.table[sound.id][volumeLevel]
-                end
-                PlaySoundFile(soundFile, "MASTER")
-            end
-            TrackedFactions[faction] = newstanding
-        end
+    if debugMode then
+        self:DebugMessage("[|cff3bbc00RU|r] [|cff808080DEBUG|r] |cffc586c0ReputationRankIncrease|r event triggered for |cff00e012" .. factionName .. "|r (Rank: " .. newRank .. ")")
     end
-end
---===============================================
-function TestRepSound()
     local sound = SelectSound(RU.db.profile.RepSoundSelect)
     local volumeLevel = RU.db.profile.RepVolume
-    local soundFile
-    if volumeLevel == 0 then
-        return
-    end
-    if sound.id == 1 then
-        soundFile = defaultSounds[6][volumeLevel]
-    else
-        soundFile = sound.table[sound.id][volumeLevel]
-    end
-    if soundFile then
-        PlaySoundFile(soundFile, "MASTER")
+    PlaySelectedSound(sound, volumeLevel, defaultSounds[6])
+end
+
+--=====================================================================================
+-- ChatFrame Hooks
+--=====================================================================================
+function RU:ReputationChatFrameHook()
+    if not chatFrameHooked then
+        ChatFrame_AddMessageEventFilter("CHAT_MSG_SYSTEM", function(self, event, msg)
+            for _, rank in ipairs(reputationRanks) do
+                local reputationGainPattern = "You are now " .. rank .. " with (.+)%.?"
+                local factionName = string.match(msg, reputationGainPattern)
+                if factionName then
+                    RU:ReputationRankIncrease(factionName, rank)
+                end
+            end
+            return false -- Ensure the original message is not blocked
+        end)
+        chatFrameHooked = true
     end
 end
+
 --=====================================================================================
+-- Test Functions
 --=====================================================================================
-function RU:SlashCommand(input, editbox)
-    if input == "enable" then
-        self:Enable()
-        self:Print("Enabled.")
-    elseif input == "disable" then
-        self:Disable()
-        self:Print("Disabled.")
-    else
-        InterfaceOptionsFrame_OpenToCategory(self.optionsFrame)
-        InterfaceOptionsFrame_OpenToCategory(self.optionsFrame)
+function RU:TestRenownSound()
+    if debugMode then
+        self:DebugMessage("[|cff3bbc00RU|r] [|cff808080DEBUG|r] |cffc586c0TestRenownSound|r event triggered.")
     end
+    local sound = SelectSound(RU.db.profile.RenownSoundSelect)
+    local volumeLevel = RU.db.profile.RenownVolume
+    PlaySelectedSound(sound, volumeLevel, defaultSounds[5])
+end
+
+function RU:TestRepSound()
+    if debugMode then
+        self:DebugMessage("[|cff3bbc00RU|r] [|cff808080DEBUG|r] |cffc586c0TestRepSound|r event triggered.")
+    end
+    local sound = SelectSound(self.db.profile.RepSoundSelect)
+    local volumeLevel = self.db.profile.RepVolume
+    PlaySelectedSound(sound, volumeLevel, defaultSounds[6])
 end