+++ data/sounds.lua
@@ -1,35 +1,83 @@
 --=====================================================================================
+-- Sound IDs
 --=====================================================================================
+
+-- Retail
 muteSoundIDs = {
-    569143,
-    1489546,
-    569593, -- Level Up
-    642841,
-    4745441,
-    568016,
-    567400, -- Quest Accepted
-    567439, -- Quest Turned In
-    2066672
+    4745441, -- Renown
+    568016,  -- Reputation (8473)
 }
---=====================================================================================
---=====================================================================================
-muteSoundIDs_c = {
-    569143,
-    569593, -- Level Up
-    642841,
-    568016,
-    567400, -- Quest Accepted
-    567439  -- Quest Turned In
-}
---=====================================================================================
---=====================================================================================
+
+-- Vanilla
 muteSoundIDs_v = {
-    569593, -- Level Up
-    568016,
-    567400, -- Quest Accepted
-    567439  -- Quest Turned In
+    568016,  -- Reputation (8473)
 }
+
 --=====================================================================================
+-- Sound Options
+--=====================================================================================
+soundOptions = {
+    "[Default]",
+    "[Random]",
+    "Altered Beast",
+    "Assassin's Creed",
+    "Castlevania",
+    "Diablo 2",
+    "Dragon Quest",
+    "DotA 2",
+    "Elden Ring (1)",
+    "Elden Ring (2)",
+    "Elden Ring (3)",
+    "Elden Ring (4)",
+    "Elden Ring (5)",
+    "Elden Ring (6)",
+    "EverQuest",
+    "Fallout - New Vegas",
+    "Fallout 3",
+    "Final Fantasy",
+    "Fire Emblem",
+    "Fire Emblem - Awakening",
+    "Fly For Fun",
+    "Fortnite",
+    "GTA - San Andreas",
+    "Kingdom Hearts 3",
+    "Kirby (1)",
+    "Kirby (2)",
+    "League of Legends",
+    "Legend of Zelda",
+    "Maplestory",
+    "Metal Gear Solid",
+    "Minecraft",
+    "Modern Warfare 2",
+    "Morrowind",
+    "Old School Runescape",
+    "Palworld",
+    "Path of Exile",
+    "Pokemon",
+    "Ragnarok Online",
+    "Shining Force II",
+    "Shining Force III (1)",
+    "Shining Force III (2)",
+    "Shining Force III (3)",
+    "Shining Force III (4)",
+    "Shining Force III (5)",
+    "Shining Force III (6)",
+    "Shining Force III (7)",
+    "Shining Force III (8)",
+    "Shining Force III (9)",
+    "Shining Force III (10)",
+    "Shining Force III (11)",
+    "Skyrim",
+    "Sonic The Hedgehog",
+    "Spyro The Dragon",
+    "Super Mario Bros 3",
+    "Warcraft 3",
+    "Witcher 3 (1)",
+    "Witcher 3 (2)",
+}
+
+--=====================================================================================
+-- Default Sounds
 --=====================================================================================
 defaultSounds = {
     [1] = { -- Achievement
@@ -78,7 +126,9 @@ defaultSounds = {
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\post_default_high.ogg"
     }
 }
+
 --=====================================================================================
+-- Custom Sounds
 --=====================================================================================
 sounds = {
     [3] = {
@@ -112,220 +162,248 @@ sounds = {
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\dota_2_high.ogg"
     },
     [9] = {
+        [1] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-1_low.ogg",
+        [2] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-1_med.ogg",
+        [3] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-1_high.ogg"
+    },
+    [10] = {
+        [1] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-2_low.ogg",
+        [2] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-2_med.ogg",
+        [3] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-2_high.ogg"
+    },
+    [11] = {
+        [1] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-3_low.ogg",
+        [2] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-3_med.ogg",
+        [3] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-3_high.ogg"
+    },
+    [12] = {
+        [1] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-4_low.ogg",
+        [2] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-4_med.ogg",
+        [3] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-4_high.ogg"
+    },
+    [13] = {
+        [1] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-5_low.ogg",
+        [2] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-5_med.ogg",
+        [3] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-5_high.ogg"
+    },
+    [14] = {
+        [1] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-6_low.ogg",
+        [2] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-6_med.ogg",
+        [3] = "Interface\\Addons\\Rep Up!\\sounds\\elden_ring-6_high.ogg"
+    },
+    [15] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\everquest_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\everquest_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\everquest_high.ogg"
     },
-    [10] = {
+    [16] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\fallout_new_vegas_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\fallout_new_vegas_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\fallout_new_vegas_high.ogg"
     },
-    [11] = {
+    [17] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\fallout_3_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\fallout_3_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\fallout_3_high.ogg"
     },
-    [12] = {
+    [18] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\final_fantasy_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\final_fantasy_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\final_fantasy_high.ogg"
     },
-    [13] = {
+    [19] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\fire_emblem_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\fire_emblem_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\fire_emblem_high.ogg"
     },
-    [14] = {
+    [20] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\fire_emblem_awakening_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\fire_emblem_awakening_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\fire_emblem_awakening_high.ogg"
     },
-    [15] = {
+    [21] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\fly_for_fun_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\fly_for_fun_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\fly_for_fun_high.ogg"
     },
-    [16] = {
+    [22] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\fortnite_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\fortnite_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\fortnite_high.ogg"
     },
-    [17] = {
+    [23] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\gta_san_andreas_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\gta_san_andreas_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\gta_san_andreas_high.ogg"
     },
-    [18] = {
+    [24] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\kingdom_hearts_3_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\kingdom_hearts_3_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\kingdom_hearts_3_high.ogg"
     },
-    [19] = {
+    [25] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\kirby-1_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\kirby-1_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\kirby-1_high.ogg"
     },
-    [20] = {
+    [26] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\kirby-2_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\kirby-2_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\kirby-2_high.ogg"
     },
-    [21] = {
+    [27] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\league_of_legends_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\league_of_legends_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\league_of_legends_high.ogg"
     },
-    [22] = {
+    [28] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\legend_of_zelda_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\legend_of_zelda_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\legend_of_zelda_high.ogg"
     },
-    [23] = {
+    [29] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\maplestory_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\maplestory_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\maplestory_high.ogg"
     },
-    [24] = {
+    [30] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\metalgear_solid_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\metalgear_solid_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\metalgear_solid_high.ogg"
     },
-    [25] = {
+    [31] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\minecraft_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\minecraft_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\minecraft_high.ogg"
     },
-    [26] = {
+    [32] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\modern_warfare_2_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\modern_warfare_2_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\modern_warfare_2_high.ogg"
     },
-    [27] = {
+    [33] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\morrowind_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\morrowind_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\morrowind_high.ogg"
     },
-    [28] = {
+    [34] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\old_school_runescape_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\old_school_runescape_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\old_school_runescape_high.ogg"
     },
-    [29] = {
+    [35] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\palworld_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\palworld_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\palworld_high.ogg"
     },
-    [30] = {
+    [36] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\path_of_exile_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\path_of_exile_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\path_of_exile_high.ogg"
     },
-    [31] = {
+    [37] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\pokemon_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\pokemon_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\pokemon_high.ogg"
     },
-    [32] = {
+    [38] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\ragnarok_online_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\ragnarok_online_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\ragnarok_online_high.ogg"
     },
-    [33] = {
+    [39] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_2_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_2_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_2_high.ogg"
     },
-    [34] = {
+    [40] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-1_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-1_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-1_high.ogg"
     },
-    [35] = {
+    [41] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-2_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-2_med.ogg",
-        [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-2_haigh.ogg"
+        [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-2_high.ogg"
     },
-    [36] = {
+    [42] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-3_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-3_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-3_high.ogg"
     },
-    [37] = {
+    [43] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-4_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-4_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-4_high.ogg"
     },
-    [38] = {
+    [44] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-5_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-5_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-5_high.ogg"
     },
-    [39] = {
+    [45] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-6_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-6_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-6_high.ogg"
     },
-    [40] = {
+    [46] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-7_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-7_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-7_high.ogg"
     },
-    [41] = {
+    [47] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-8_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-8_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-8_high.ogg"
     },
-    [42] = {
+    [48] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-9_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-9_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-9_high.ogg"
     },
-    [43] = {
+    [49] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-10_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-10_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-10_high.ogg"
     },
-    [44] = {
+    [50] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-11_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-11_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\shining_force_3-11_high.ogg"
     },
-    [45] = {
+    [51] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\skyrim_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\skyrim_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\skyrim_high.ogg"
     },
-    [46] = {
+    [52] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\sonic_the_hedgehog_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\sonic_the_hedgehog_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\sonic_the_hedgehog_high.ogg"
     },
-    [47] = {
+    [53] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\spyro_the_dragon_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\spyro_the_dragon_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\spyro_the_dragon_high.ogg"
     },
-    [48] = {
+    [54] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\super_mario_bros_3_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\super_mario_bros_3_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\super_mario_bros_3_high.ogg"
     },
-    [49] = {
+    [55] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\warcraft_3_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\warcraft_3_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\warcraft_3_high.ogg"
     },
-    [50] = {
+    [56] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\witcher_3-1_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\witcher_3-1_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\witcher_3-1_high.ogg"
     },
-    [51] = {
+    [57] = {
         [1] = "Interface\\Addons\\Rep Up!\\sounds\\witcher_3-2_low.ogg",
         [2] = "Interface\\Addons\\Rep Up!\\sounds\\witcher_3-2_med.ogg",
         [3] = "Interface\\Addons\\Rep Up!\\sounds\\witcher_3-2_high.ogg"
-    },
+    }
 }
---=====================================================================================
---=====================================================================================